+++ audionimbus/tests/simulation.rs
@@ -31,10 +31,7 @@ fn test_simulation() {
     let scene = Scene::try_new(&context).unwrap();
     simulator.set_scene(&scene);
 
-    let source_settings = SourceSettings {
-        flags: SimulationFlags::DIRECT | SimulationFlags::REFLECTIONS,
-    };
-    let source = Source::try_new(&simulator, source_settings).unwrap();
+    let source = Source::try_new(&simulator).unwrap();
 
     let pathing_probes = ProbeBatch::try_new(&context).unwrap();
     let simulation_inputs = SimulationInputs::new(CoordinateSystem {
@@ -69,9 +66,7 @@ fn test_simulation() {
         find_alternate_paths: true,
         deviation: DeviationModel::default(),
     });
-    source
-        .set_inputs(SimulationFlags::DIRECT, &simulation_inputs)
-        .unwrap();
+    source.set_direct_inputs(&simulation_inputs).unwrap();
 
     simulator.add_source(&source);
 
@@ -84,10 +79,7 @@ fn test_simulation() {
             irradiance_min_distance: 1.0,
         });
     simulator
-        .set_shared_inputs(
-            SimulationFlags::DIRECT | SimulationFlags::REFLECTIONS,
-            &simulation_shared_inputs,
-        )
+        .set_shared_inputs(&simulation_shared_inputs)
         .unwrap();
 
     simulator.commit();
@@ -215,10 +207,7 @@ fn test_pathing_without_probes() {
         .bake(&context, &mut pathing_probes, &scene, path_bake_params)
         .unwrap();
 
-    let source_settings = SourceSettings {
-        flags: SimulationFlags::PATHING,
-    };
-    let source = Source::try_new(&simulator, source_settings).unwrap();
+    let source = Source::try_new(&simulator).unwrap();
     let simulation_inputs = SimulationInputs::new(CoordinateSystem::default())
         .with_direct(
             DirectSimulationParameters::new()
@@ -246,9 +235,7 @@ fn test_pathing_without_probes() {
             find_alternate_paths: true,
             deviation: DeviationModel::default(),
         });
-    source
-        .set_inputs(SimulationFlags::PATHING, &simulation_inputs)
-        .unwrap();
+    source.set_pathing_inputs(&simulation_inputs).unwrap();
     simulator.add_source(&source);
 
     simulator.commit();
@@ -284,7 +271,7 @@ fn test_reflections_without_scene() {
             irradiance_min_distance: 1.0,
         });
     simulator
-        .set_shared_inputs(SimulationFlags::REFLECTIONS, &simulation_shared_inputs)
+        .set_shared_reflections_inputs(&simulation_shared_inputs)
         .unwrap();
 
     simulator.commit();